This is a port of https://github.com/oltpbenchmark/oltpbench/pull/320 which was written by @bglasber

The original PR description is copied below.

---

When generating an auctionmark database with a larger scale factor (e.g. 50), there are collisions in UserIds because there not enough bits allocated for OFFSET. There is some space left over in the CompositeId, so I've allocated a few more bits. Also, I've reordered the fields in UserId to match how they are packed in the compositeId. I've verified that this fix works to generate an auctionmark database with scalefactor 50.

The number of rows also exceeds the max values for short, so I've switched those to int in AuctionmarkLoader.

I propose changing the encode method on CompositeId to throw an IllegalArgumentException rather than assert. By default, asserts are not triggered, which means that we receive no warning that IDs may collide. This issue results in confusing errors that are difficult to track down for end-users. I've added code and a test to this effect.

All other tests pass with ant junit.

+++ src/main/java/com/oltpbenchmark/benchmarks/auctionmark/util/LoaderItemInfo.java
@@ -35,7 +35,7 @@ public class LoaderItemInfo extends ItemInfo {
     private short numImages;
     private short numAttributes;
     private short numComments;
-    private short numWatches;
+    private int numWatches;
     private Timestamp startDate;
     private Timestamp purchaseDate;
     private float initialPrice;
@@ -78,11 +78,11 @@ public class LoaderItemInfo extends ItemInfo {
         this.numComments = numComments;
     }
 
-    public short getNumWatches() {
+    public int getNumWatches() {
         return numWatches;
     }
 
-    public void setNumWatches(short numWatches) {
+    public void setNumWatches(int numWatches) {
         this.numWatches = numWatches;
     }
